abstract ActiveInteractor::Base into composable modules
abstract `ActiveInteractor::Base` into composable modules to make future expansion easier.

- added `ActiveInteractor::ActiveModelErrorMethods`
- added `ActiveInteractor::Context::Result.register_owner`
- added `ActiveInteractor::Base.input_type`
- added `ActiveInteractor::Base.output_type`
- added `ActiveInteractor::Interactor`
- added `ActiveInteractor::Interactor::Base`
- added `ActiveInteractor::Interactor::ContextMethods`
- added `ActiveInteractor::Interactor::InteractionMethods`
- added `ActiveInteractor::Type::DeclerationMethods`
- changed `ActiveInteractor::Context::Result.for_output_context` will no longer set the `::ResultContext` constant on the
  interactor.  Use `ActiveInteractor::Context::Result.register_owner` to do this
- removed `ActiveInteractor::HasActiveModelErrors` use `ActiveInteractor::ActiveModelErrorMethods` instead
- removed `ActiveInteractor::Type::HasTypes` use `ActiveInteractor::Type::DeclerationMethods` instead
