+++ lib/active_interactor/context/result.rb
@@ -3,8 +3,11 @@
 module ActiveInteractor
   module Context
     class Result
-      def self.for_output_context(owner, context)
+      def self.register_owner(owner)
         owner.const_set(:ResultContext, Class.new(self))
+      end
+
+      def self.for_output_context(owner, context)
         context.fields.each_key { |field| owner::ResultContext.send(:attr_reader, field) }
         owner::ResultContext.new(context.fields)
       end
